Move nospam from Settings to Profile
