+++ app/src/main/res/xml/pref_profile.xml
@@ -59,6 +59,10 @@
         android:textIsSelectable="true"
         android:title="@string/pref_title_tox_id" />
 
+    <Preference
+        android:key="nospam"
+        android:title="@string/pref_nospam" />
+
     <CheckBoxPreference
         android:defaultValue="true"
         android:key="logging_enabled"
